Add files via upload
test.py is to reset the database when needed. 
the slides are provided as a form of aid to explain individual functions we've done.
venv is required to open this file, import of external programs is listed in a docx. 
Templates credit to Colorlib( footer, error page), and the bootstrap made 4 (Home Page)